+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project (not yet) adheres to [Semantic Versioning](https://semver.org/spec/
 
 - Password visibility toggle added to login and signup forms
 - External projects/plans show icon to indicate external link
+- Offline Event is now A Module
 
 ### Fixed